Cleanup.

+++ src/level3/dgemm_nn.c
@@ -1,7 +1,5 @@
 #include <ulmblas.h>
 #include <stdio.h>
-#include <emmintrin.h>
-#include <immintrin.h>
 
 #define MC  384
 #define KC  384
@@ -119,10 +117,10 @@ pack_B(int kc, int nc, const double *B, int incRowB, int incColB,
 //  Micro kernel for multiplying panels from A and B.
 //
 static void
-dgemm_micro_kernel(long kc,
+dgemm_micro_kernel(int kc,
                    double alpha, const double *A, const double *B,
                    double beta,
-                   double *C, long incRowC, long incColC)
+                   double *C, int incRowC, int incColC)
 {
     double AB[MR*NR];
